refactor(collection): do not use resultMutator for insertMany
Fixes NODE-1895
