+++ lib/operations/collection_ops.js
@@ -893,6 +893,63 @@ function insertOne(coll, doc, options, callback) {
 }
 
 /**
+ * Inserts an array of documents into MongoDB. If documents passed in do not contain the **_id** field,
+ * one will be added to each of the documents missing it by the driver, mutating the document. This behavior
+ * can be overridden by setting the **forceServerObjectId** flag.
+ *
+ * @method
+ * @param {Collection} a Collection instance.
+ * @param {object[]} docs Documents to insert.
+ * @param {number} [options] Optional settings. See Collection.prototype.insertMany for a list of options.
+ * @param {Collection~insertWriteOpCallback} [callback] The command result callback
+ */
+function insertMany(coll, docs, options, callback) {
+  if (!Array.isArray(docs) && typeof callback === 'function') {
+    return callback(
+      MongoError.create({ message: 'docs parameter must be an array of documents', driver: true })
+    );
+  } else if (!Array.isArray(docs)) {
+    return new coll.s.promiseLibrary((resolve, reject) => {
+      reject(
+        MongoError.create({ message: 'docs parameter must be an array of documents', driver: true })
+      );
+    });
+  }
+
+  // If keep going set unordered
+  options['serializeFunctions'] = options['serializeFunctions'] || coll.s.serializeFunctions;
+
+  docs = prepareDocs(coll, docs, options);
+
+  // Generate the bulk write operations
+  const operations = [
+    {
+      insertMany: docs
+    }
+  ];
+
+  bulkWrite(coll, operations, options, (err, result) => {
+    if (err) return callback(err, null);
+    callback(null, mapInsertManyResults(docs, result));
+  });
+}
+
+function mapInsertManyResults(docs, r) {
+  const finalResult = {
+    result: { ok: 1, n: r.insertedCount },
+    ops: docs,
+    insertedCount: r.insertedCount,
+    insertedIds: r.insertedIds
+  };
+
+  if (r.getLastOp()) {
+    finalResult.result.opTime = r.getLastOp();
+  }
+
+  return finalResult;
+}
+
+/**
  * Determine whether the collection is a capped collection.
  *
  * @method
@@ -1476,6 +1533,7 @@ module.exports = {
   indexes,
   indexExists,
   indexInformation,
+  insertMany,
   insertOne,
   isCapped,
   mapReduce,
