+++ lib/utils.js
@@ -360,7 +360,6 @@ var mergeOptionsAndWriteConcern = function(targetOptions, sourceOptions, keys, m
  * @param {function} operation The operation to execute
  * @param {array} args Arguments to apply the provided operation
  * @param {object} [options] Options that modify the behavior of the method
- * @param {function]} [options.resultMutator] Allows for the result of the operation to be changed for custom return types
  */
 const executeOperation = (topology, operation, args, options) => {
   if (topology == null) {
@@ -373,7 +372,6 @@ const executeOperation = (topology, operation, args, options) => {
 
   options = options || {};
   const Promise = topology.s.promiseLibrary;
-  let resultMutator = options.resultMutator;
   let callback = args[args.length - 1];
 
   // The driver sessions spec mandates that we implicitly create sessions for operations
@@ -397,12 +395,10 @@ const executeOperation = (topology, operation, args, options) => {
         session.endSession(() => {
           delete opOptions.session;
           if (err) return reject(err);
-          if (resultMutator) return resolve(resultMutator(result));
           resolve(result);
         });
       } else {
         if (err) return reject(err);
-        if (resultMutator) return resolve(resultMutator(result));
         resolve(result);
       }
     };
